SISRP-5405 Un-package Email classes
