--- app/models/campus_solutions/models/my_email.rb
@@ -1,12 +1,0 @@
-module CampusSolutions
-  module Models
-    class MyEmail < UserSpecificModel
-
-      def update(params = {})
-        proxy = CampusSolutions::Posts::Email.new({user_id: @uid, params: params})
-        proxy.get
-      end
-
-    end
-  end
-end
