+++ app/models/campus_solutions/my_email.rb
@@ -1,0 +1,10 @@
+module CampusSolutions
+  class MyEmail < UserSpecificModel
+
+    def update(params = {})
+      proxy = CampusSolutions::Email.new({user_id: @uid, params: params})
+      proxy.get
+    end
+
+  end
+end
